Ladder - step3 - test (잔여 테스트코드 수정)
